+++ include/Point.h
@@ -1,6 +1,5 @@
-#include <GL/glut.h>
 
 struct Point
 {
-    GLfloat x, y;
+    double x, y;
 };
