+++ src/BaseEntity.cpp
@@ -1,6 +1,6 @@
 #include "../include/BaseEntity.h"
 
-void BaseEntity::setCoords(float *_coords)
+void BaseEntity::setCoords(double *_coords)
 {
 	for (int i = 0; i < 4; ++i)
 	{
